add a docker image configuration to extract ROS model from existing ROS code
